🚚 api: move usecase to src context
