--- api/lib/domain/usecases/get-session-results.js
@@ -1,8 +1,0 @@
-const getSessionResults = async function ({ sessionId, sessionEnrolmentRepository, certificationResultRepository }) {
-  const session = await sessionEnrolmentRepository.get({ id: sessionId });
-  const certificationResults = await certificationResultRepository.findBySessionId({ sessionId });
-
-  return { session, certificationResults };
-};
-
-export { getSessionResults };
