+++ api/tests/certification/results/unit/domain/usecases/get-session-results_test.js
@@ -1,5 +1,5 @@
-import { getSessionResults } from '../../../../lib/domain/usecases/get-session-results.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { getSessionResults } from '../../../../../../src/certification/results/domain/usecases/get-session-results.js';
+import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Use Cases | get-session-results', function () {
   const sessionEnrolmentRepository = { get: null };
